Move subroutines from server_show to helper_show
